+++ nym-vpn-core/crates/nym-vpn-account-controller/src/command_sender.rs
@@ -17,9 +17,10 @@ use nym_vpn_lib_types::{
 };
 use nym_vpn_store::types::StorableAccount;
 use tokio::sync::mpsc::UnboundedSender;
+use tracing::instrument;
 use url::Url;
 
-#[derive(Clone)]
+#[derive(Clone, Debug)]
 pub struct AccountCommandSender {
     command_tx: UnboundedSender<AccountCommand>,
 }
@@ -31,6 +32,7 @@ impl AccountCommandSender {
         Self { command_tx }
     }
 
+    #[instrument]
     pub async fn store_account(&self, account: StorableAccount) -> Result<(), AccountCommandError> {
         let (tx, rx) = ReturnSender::new();
         self.command_tx
@@ -39,6 +41,7 @@ impl AccountCommandSender {
         rx.await.map_err(AccountCommandError::internal)?
     }
 
+    #[instrument]
     pub async fn register_account(
         &self,
         account: StorableAccount,
@@ -51,6 +54,7 @@ impl AccountCommandSender {
         rx.await.map_err(AccountCommandError::internal)?
     }
 
+    #[instrument]
     pub async fn get_stored_account(&self) -> Result<Option<StorableAccount>, AccountCommandError> {
         let (tx, rx) = ReturnSender::new();
         self.command_tx
@@ -59,6 +63,7 @@ impl AccountCommandSender {
         rx.await.map_err(AccountCommandError::internal)?
     }
 
+    #[instrument]
     pub async fn create_account_command(&self) -> Result<(), AccountCommandError> {
         let (tx, rx) = ReturnSender::new();
         self.command_tx
@@ -68,6 +73,7 @@ impl AccountCommandSender {
         Ok(())
     }
 
+    #[instrument]
     pub async fn get_account_id(&self) -> Result<Option<String>, AccountCommandError> {
         let (tx, rx) = ReturnSender::new();
         self.command_tx
@@ -78,6 +84,7 @@ impl AccountCommandSender {
         rx.await.map_err(AccountCommandError::internal)?
     }
 
+    #[instrument]
     pub async fn forget_account(&self) -> Result<(), AccountCommandError> {
         let (tx, rx) = ReturnSender::new();
         self.command_tx
@@ -86,6 +93,7 @@ impl AccountCommandSender {
         rx.await.map_err(AccountCommandError::internal)?
     }
 
+    #[instrument]
     pub async fn rotate_keys(&self) -> Result<(), AccountCommandError> {
         let (tx, rx) = ReturnSender::new();
         self.command_tx
@@ -94,6 +102,7 @@ impl AccountCommandSender {
         rx.await.map_err(AccountCommandError::internal)?
     }
 
+    #[instrument]
     pub async fn background_refresh_account_state(&self) -> Result<(), AccountCommandError> {
         let (tx, rx) = ReturnSender::new();
         self.command_tx
@@ -102,6 +111,7 @@ impl AccountCommandSender {
         rx.await.map_err(AccountCommandError::internal)?
     }
 
+    #[instrument]
     pub async fn reset_device_identity(
         &self,
         seed: Option<[u8; 32]>,
@@ -113,6 +123,7 @@ impl AccountCommandSender {
         rx.await.map_err(AccountCommandError::internal)?
     }
 
+    #[instrument]
     pub async fn get_usage(&self) -> Result<Vec<NymVpnUsage>, AccountCommandError> {
         let (tx, rx) = ReturnSender::new();
         self.command_tx
@@ -121,6 +132,7 @@ impl AccountCommandSender {
         rx.await.map_err(AccountCommandError::internal)?
     }
 
+    #[instrument]
     pub async fn get_device_identity(&self) -> Result<Option<String>, AccountCommandError> {
         let (tx, rx) = ReturnSender::new();
         self.command_tx
@@ -129,6 +141,7 @@ impl AccountCommandSender {
         rx.await.map_err(AccountCommandError::internal)?
     }
 
+    #[instrument]
     pub async fn get_devices(&self) -> Result<Vec<NymVpnDevice>, AccountCommandError> {
         let (tx, rx) = ReturnSender::new();
         self.command_tx
@@ -137,6 +150,7 @@ impl AccountCommandSender {
         rx.await.map_err(AccountCommandError::internal)?
     }
 
+    #[instrument]
     pub async fn get_active_devices(&self) -> Result<Vec<NymVpnDevice>, AccountCommandError> {
         let (tx, rx) = ReturnSender::new();
         self.command_tx
@@ -145,6 +159,7 @@ impl AccountCommandSender {
         rx.await.map_err(AccountCommandError::internal)?
     }
 
+    #[instrument]
     pub async fn get_available_tickets(&self) -> Result<AvailableTicketbooks, AccountCommandError> {
         let (tx, rx) = ReturnSender::new();
         self.command_tx
@@ -155,6 +170,7 @@ impl AccountCommandSender {
         rx.await.map_err(AccountCommandError::internal)?
     }
 
+    #[instrument]
     pub async fn get_account_summary(
         &self,
     ) -> Result<Option<VpnAccountSummary>, AccountCommandError> {
@@ -165,6 +181,7 @@ impl AccountCommandSender {
         rx.await.map_err(AccountCommandError::internal)?
     }
 
+    #[instrument]
     pub async fn get_deeplink(
         &self,
         kind: DeeplinkKind,
@@ -185,6 +202,7 @@ impl AccountCommandSender {
         rx.await.map_err(AccountCommandError::internal)?
     }
 
+    #[instrument]
     pub async fn set_resolver_overrides(
         &self,
         resolver_overrides: Option<ResolverOverrides>,
@@ -199,6 +217,7 @@ impl AccountCommandSender {
         rx.await.map_err(AccountCommandError::internal)?
     }
 
+    #[instrument]
     pub async fn set_vpn_api_firewall_up(&self) -> Result<(), AccountCommandError> {
         let (tx, rx) = ReturnSender::new();
         self.command_tx
@@ -207,6 +226,7 @@ impl AccountCommandSender {
         rx.await.map_err(AccountCommandError::internal)?
     }
 
+    #[instrument]
     pub async fn set_vpn_api_firewall_down(&self) -> Result<(), AccountCommandError> {
         let (tx, rx) = ReturnSender::new();
         self.command_tx
@@ -215,6 +235,7 @@ impl AccountCommandSender {
         rx.await.map_err(AccountCommandError::internal)?
     }
 
+    #[instrument]
     pub async fn decentralised_balance(&self) -> Result<Vec<Coin>, AccountCommandError> {
         let (tx, rx) = ReturnSender::new();
         self.command_tx
@@ -223,6 +244,7 @@ impl AccountCommandSender {
         rx.await.map_err(AccountCommandError::internal)?
     }
 
+    #[instrument]
     pub async fn decentralised_obtain_ticketbooks(
         &self,
         amount: u64,
@@ -234,6 +256,7 @@ impl AccountCommandSender {
         rx.await.map_err(AccountCommandError::internal)?
     }
 
+    #[instrument]
     pub async fn query_upgrade_mode_enabled(&self) -> Result<bool, AccountCommandError> {
         let (tx, rx) = ReturnSender::new();
         self.command_tx
@@ -244,6 +267,7 @@ impl AccountCommandSender {
         rx.await.map_err(AccountCommandError::internal)?
     }
 
+    #[instrument]
     pub async fn send_disable_upgrade_mode(&self) -> Result<(), AccountCommandError> {
         let (tx, rx) = ReturnSender::new();
         self.command_tx